+++ src/plugins/background-geolocation.ts
@@ -365,14 +365,12 @@ export class BackgroundGeolocation {
    * @param {Function} callback callback will be called when background location is determined.
    * @param {Function} errorCallback callback to be executed every time a geolocation error occurs.
    * @param {Config} options An object of type Config
-   * @return Location object, which tries to mimic w3c Coordinates interface.
-   * See http://dev.w3.org/geo/api/spec-source.html#coordinates_interface
-   * Callback to be executed every time a geolocation is recorded in the background.
+   * @return {Promise<any>}
    */
   @Cordova({
-    sync: true
+    callbackOrder: 'reverse'
   })
-  static configure(callback: Function, errorCallback: Function, options: BackgroundGeolocationConfig): any { return; }
+  static configure(options: BackgroundGeolocationConfig): Promise<any> { return; }
 
   /**
    * Turn ON the background-geolocation system.
@@ -391,17 +389,21 @@ export class BackgroundGeolocation {
 
   /**
    * Inform the native plugin that you're finished, the background-task may be completed
-   * NOTE: IOS, WP only
+   * @returns {Promise<any>}
    */
-  @Cordova()
-  static finish() { }
+  @Cordova({
+    platforms: ['iOS', 'Windows Phone']
+  })
+  static finish(): Promise<any> { return; }
 
   /**
    * Force the plugin to enter "moving" or "stationary" state
-   * NOTE: IOS, WP only
+   * @returns {Promise<any>}
    */
-  @Cordova()
-  static changePace(isMoving: boolean) { }
+  @Cordova({
+    platforms: ['iOS', 'Windows Phone']
+  })
+  static changePace(isMoving: boolean): Promise<any> { return; }
 
   /**
    * Setup configuration
@@ -414,27 +416,30 @@ export class BackgroundGeolocation {
 
   /**
    * Returns current stationaryLocation if available. null if not
-   * NOTE: IOS, WP only
    * @returns {Promise<Location>}
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['iOS', 'Windows Phone']
+  })
   static getStationaryLocation(): Promise<BackgroundGeolocationResponse> { return; }
 
   /**
    * Add a stationary-region listener. Whenever the devices enters "stationary-mode",
    * your #success callback will be executed with #location param containing #radius of region
-   * NOTE: IOS, WP only
    * @returns {Promise<any>}
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['iOS', 'Windows Phone']
+  })
   static onStationary(): Promise<any> { return; }
 
   /**
    * Check if location is enabled on the device
    * @returns {Promise<number>} Returns a promise with int argument that takes values 0, 1 (true).
-   * NOTE: ANDROID only
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['Android']
+  })
   static isLocationEnabled(): Promise<number> { return; }
 
   /**
@@ -453,18 +458,21 @@ export class BackgroundGeolocation {
    * Method can be used to detect user changes in location services settings.
    * If user enable or disable location services then success callback will be executed.
    * In case or error (SettingNotFoundException) fail callback will be executed.
-   * NOTE: ANDROID only
    * @returns {Promise<boolean>}
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['Android']
+  })
   static watchLocationMode(): Promise<boolean> { return; }
 
   /**
    * Stop watching for location mode changes.
-   * NOTE: ANDROID only
+   * @returns {Promise<any>}
    */
-  @Cordova()
-  static stopWatchingLocationMode() { }
+  @Cordova({
+    platforms: ['Android']
+  })
+  static stopWatchingLocationMode(): Promise<any> { return; }
 
   /**
    * Method will return all stored locations.
@@ -473,10 +481,11 @@ export class BackgroundGeolocation {
    *    by the system
    *  or
    *  - option.debug is true
-   * NOTE: ANDROID only
    * @returns {Promise<any>}
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['Android']
+  })
   static getLocations(): Promise<any> { return; }
 
   /** 
@@ -488,18 +497,20 @@ export class BackgroundGeolocation {
 
   /**
    * Delete stored location by given locationId.
-   * NOTE: ANDROID only
    * @returns {Promise<any>}
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['Android']
+  })
   static deleteLocation(locationId: number): Promise<any> { return; }
 
   /**
    * Delete all stored locations.
-   * NOTE: ANDROID only
    * @returns {Promise<any>}
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['Android']
+  })
   static deleteAllLocations(): Promise<any> { return; }
 
   /**
@@ -511,13 +522,13 @@ export class BackgroundGeolocation {
    *
    * BackgroundGeolocation.Mode.FOREGROUND
    * BackgroundGeolocation.Mode.BACKGROUND 
-   *
-   * NOTE: iOS only
-   *
+   **
    * @param {number} See above. 
    * @returns {Promise<any>}
    */
-  @Cordova()
+  @Cordova({
+    platforms: ['iOS']
+  })
   static switchMode(modeId: number): Promise<any> { return; }
 
   /**