+++ src/plugins/localnotifications.ts
@@ -65,6 +65,13 @@ export interface ILocalNotification {
    */
   smallIcon?: string;
 
+  /**
+  * ANDROID ONLY
+  * RGB value for the background color of the smallIcon.
+  * Default: Androids COLOR_DEFAULT, which will vary based on Android version.
+  */
+  color?: string;
+
 
   /**
    * ANDROID ONLY