+++ src/plugins/media.ts
@@ -45,9 +45,7 @@ export interface MediaError {
  * });
  *
  * // get file duration
- * file.getDuration().then((duration) => {
- *   console.log(position);
- * });
+ * file.getDuration();
  *
  * // skip to 10 seconds (expects int value in ms)
  * file.seekTo(10000);
@@ -227,6 +225,21 @@ export class MediaPlugin {
   })
   stopRecord(): void { }
 
+  /**
+   * Pauses recording
+   */
+  @CordovaInstance({
+    sync: true
+  })
+  pauseRecord(): void { }
+
+  /**
+   * Resumes recording
+   */
+  @CordovaInstance({
+    sync: true
+  })
+  resumeRecord(): void { }
 
   /**
    * Stops playing an audio file.